Fixed bug in search. Code simplification.
An error (Bottom Overflowed) occurred when opening the keyboard. It was fixed by adding 'resizeToAvoidBottomInset : false' to the main_page file.
The search field and the settings button were moved to a separate file called search.bar . The themes folder, in which the fons file is created, has been added for text stylization.